requirements.txt
Mastodon.py>=1.8.0
requests>=2.31.0
python-dotenv>=1.0.0
